cleanup: move test files, styling
