--- go-config.code-workspace
@@ -1,14 +1,0 @@
-{
-	"folders": [
-		{
-			"path": "."
-		},
-		{
-			"path": "../tc2-hat-controller"
-		},
-		{
-			"path": "../management-interface"
-		}
-	],
-	"settings": {}
-}
